chore(flow): move the cup transform back to babel dir
