+++ babel/transform-cup-globals.js
